python/requirements.txt
pymongo==4.8.0
requests==2.32.3
python-dotenv==1.0.1


